Refactor into MVC-packages
