+++ app/src/main/AndroidManifest.xml
@@ -20,10 +20,10 @@
         <meta-data android:name="com.google.android.gms.version"
             android:value="@integer/google_play_services_version"/>
 
-        <activity android:name=".MultiPlayerActivity"
+        <activity android:name=".Controllers.MultiPlayerActivity"
             android:label="@string/app_name"></activity>
-        <activity android:name=".SinglePlayerActivity" />
-        <activity android:name=".MenuActivity">
+        <activity android:name=".Controllers.SinglePlayerActivity" />
+        <activity android:name=".Controllers.MenuActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
