+++ app/src/main/res/layout/activity_multi_player.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg_play"
-    tools:context=".MultiPlayerActivity"
+    tools:context=".Controllers.MultiPlayerActivity"
     android:visibility="visible">
 
     <Button
