+++ app/src/main/res/layout/activity_single_player.xml
@@ -5,6 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg_play"
-    tools:context=".SinglePlayerActivity">
+    tools:context=".Controllers.SinglePlayerActivity">
 
 </android.support.constraint.ConstraintLayout>
